+++ umbra/components/addons/tcpClientUi/tcpClientUi.py
@@ -15,6 +15,11 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
 import os
@@ -269,7 +274,7 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format(
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format(
 			"address", value)
 			self.Address_lineEdit.setText(value)
 		self.__address = value
@@ -341,7 +346,7 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format(
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format(
 			"fileCommand", value)
 		self.__fileCommand = value
 
@@ -375,7 +380,7 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format(
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format(
 			"connectionEnd", value)
 		self.__connectionEnd = value
 
@@ -642,7 +647,7 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 		if not editor:
 			return False
 
-		selectedText = foundations.strings.encode(editor.getSelectedText().replace(QChar(QChar.ParagraphSeparator),
+		selectedText = foundations.strings.toString(editor.getSelectedText().replace(QChar(QChar.ParagraphSeparator),
 																			QString("\n")))
 		if not selectedText:
 			return False
@@ -662,7 +667,7 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 			return False
 
 		if self.__scriptEditor.saveFile():
-			return self.sendDataToServer(foundations.strings.encode(self.__fileCommand).format(editor.file))
+			return self.sendDataToServer(foundations.strings.toString(self.__fileCommand).format(editor.file))
 
 	def sendDataToServer(self, data, timeOut=5):
 		"""
@@ -674,11 +679,11 @@ class TCPClientUi(QWidgetComponentFactory(uiFile=COMPONENT_UI_FILE)):
 		"""
 
 		if not data.endswith(self.__connectionEnd):
-			data = "{0}{1}".format(data, foundations.strings.encode(self.__connectionEnd).decode("string_escape"))
+			data = "{0}{1}".format(data, foundations.strings.toString(self.__connectionEnd).decode("string_escape"))
 
 		connection = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 		connection.settimeout(timeOut)
-		connection.connect((foundations.strings.encode(self.__address), int(self.__port)))
+		connection.connect((foundations.strings.toString(self.__address), int(self.__port)))
 		connection.send(data)
 		self.__engine.notificationsManager.notify(
 		"{0} | Socket connection command dispatched!".format(self.__class__.__name__))