+++ umbra/components/addons/traceUi/nodes.py
@@ -16,6 +16,11 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
 from PyQt4.QtCore import Qt
@@ -51,7 +56,7 @@ class ModuleNode(umbra.ui.nodes.GraphModelNode):
 	Component Interface class **Module** node.
 	"""
 
-	__family = "ModuleNode"
+	__family = "Module"
 
 	def __init__(self,
 				module=None,
@@ -143,4 +148,4 @@ class ModuleNode(umbra.ui.nodes.GraphModelNode):
 		"""
 
 		self.traced.value = foundations.trace.isTraced(self.__module)
-		self.traced.roles[Qt.DisplayRole] = foundations.strings.encode(self.traced.value).title()
+		self.traced.roles[Qt.DisplayRole] = foundations.strings.toString(self.traced.value).title()