+++ umbra/components/factory/componentsManagerUi/nodes.py
@@ -16,6 +16,11 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
 from PyQt4.QtCore import Qt
@@ -50,7 +55,7 @@ class PathNode(umbra.ui.nodes.GraphModelNode):
 		Component Interface class Model path node.
 	"""
 
-	__family = "Category"
+	__family = "Path"
 	"""Node family. ( String )"""
 
 	def __init__(self,
@@ -202,7 +207,7 @@ class ComponentNode(umbra.ui.nodes.GraphModelNode):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format(
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format(
 			"toolTipText", value)
 		self.__toolTipText = value