+++ umbra/components/factory/scriptEditor/nodes.py
@@ -16,6 +16,11 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
 import os
@@ -60,7 +65,7 @@ class EditorNode(umbra.ui.nodes.GraphModelNode):
 	Component Interface class **Editor** node.
 	"""
 
-	__family = "EditorNode"
+	__family = "Editor"
 
 	def __init__(self,
 				editor=None,
@@ -147,7 +152,7 @@ class FileNode(umbra.ui.nodes.GraphModelNode):
 	Component Interface class **File** node.
 	"""
 
-	__family = "FileNode"
+	__family = "File"
 
 	def __init__(self,
 				path=None,
@@ -203,7 +208,7 @@ class FileNode(umbra.ui.nodes.GraphModelNode):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format("path", value)
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format("path", value)
 		self.__path = value
 
 	@path.deleter
@@ -234,7 +239,7 @@ class DirectoryNode(umbra.ui.nodes.GraphModelNode):
 	Component Interface class **Directory** node.
 	"""
 
-	__family = "DirectoryNode"
+	__family = "Directory"
 
 	def __init__(self,
 				path=None,
@@ -291,7 +296,7 @@ class DirectoryNode(umbra.ui.nodes.GraphModelNode):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format("path", value)
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format("path", value)
 			assert os.path.exists(value), "'{0}' attribute: '{1}' path doesn't exists!".format("source", value)
 		self.__path = value
 
@@ -323,7 +328,7 @@ class ProjectNode(umbra.ui.nodes.GraphModelNode):
 	Component Interface class **Project** node.
 	"""
 
-	__family = "ProjectNode"
+	__family = "Project"
 
 	def __init__(self,
 				path=None,
@@ -380,7 +385,7 @@ class ProjectNode(umbra.ui.nodes.GraphModelNode):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format("path", value)
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format("path", value)
 			assert os.path.exists(value), "'{0}' attribute: '{1}' path doesn't exists!".format("source", value)
 		self.__path = value