+++ umbra/globals/constants.py
@@ -15,6 +15,11 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
 import os
@@ -46,10 +51,10 @@ class Constants():
 	"""Package major version: '**4**' ( String )"""
 	minorVersion = "0"
 	"""Package minor version: '**0**' ( String )"""
-	changeVersion = "7"
-	"""Package change version: '**7**' ( String )"""
+	changeVersion = "8"
+	"""Package change version: '**8**' ( String )"""
 	releaseVersion = ".".join((majorVersion, minorVersion, changeVersion))
-	"""Package release version: '**1.0.7**' ( String )"""
+	"""Package release version: '**1.0.8**' ( String )"""
 
 	logger = "Umbra_Logger"
 	"""Package logger name: '**Umbra_Logger**' ( String )"""
@@ -62,7 +67,7 @@ class Constants():
 	loggingSeparators = "*" * 96
 	"""Logging separators: '*' * 96 ( String )"""
 
-	encodingFormat = "utf-8"
+	encodingCodec = "utf-8"
 	"""Default encoding format: '**utf-8**' ( String )"""
 	encodingError = "ignore"
 	"""Default encoding error behavior: '**ignore**' ( String )"""