+++ umbra/preferences.py
@@ -17,9 +17,13 @@
 """
 
 #**********************************************************************************************************************
+#***	Future imports.
+#**********************************************************************************************************************
+from __future__ import unicode_literals
+
+#**********************************************************************************************************************
 #***	External imports.
 #**********************************************************************************************************************
-import os
 from PyQt4.QtCore import QSettings
 from PyQt4.QtCore import QVariant
 
@@ -99,7 +103,7 @@ class Preferences(object):
 		"""
 
 		if value is not None:
-			assert type(value) in (str, unicode), "'{0}' attribute: '{1}' type is not 'str' or 'unicode'!".format(
+			assert type(value) is unicode, "'{0}' attribute: '{1}' type is not 'unicode'!".format(
 			"file", value)
 		self.__file = value
 
@@ -221,7 +225,8 @@ class Preferences(object):
 		:param value: Current key value to save. ( Object )
 		"""
 
-		LOGGER.debug("> Saving '{0}' in '{1}' section with value: '{2}' in settings file.".format(key, section, value))
+		LOGGER.debug("> Saving '{0}' in '{1}' section with value: '{2}' in settings file.".format(
+		key, section, foundations.strings.toString(value)))
 
 		self.__settings.beginGroup(section)
 		self.__settings.setValue(key , QVariant(value))
@@ -244,6 +249,22 @@ class Preferences(object):
 		self.__settings.endGroup()
 
 		return value
+
+	def keyExists(self, section, key):
+		"""
+		This method checks if given key exists.
+
+		:param section: Current section to check key in. ( String )
+		:param key: Current key to check. ( String )
+		:return: Key existence. ( Boolean )
+		"""
+
+		LOGGER.debug("> Checking '{0}' key existence in '{1}' section.".format(key, section))
+
+		self.__settings.beginGroup(section)
+		exists = self.__settings.contains(key)
+		self.__settings.endGroup()
+		return exists
 
 	def __getDefaultSettings(self):
 		"""